F-04 Parse car plates
