+++ src/components/car/carPlates/index.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 
 import './styles.scss'
 import { ICar } from '../../../interfaces/carsInterfaces'
+import { parseCarNumber } from '../../../utils/carsUtils'
 
 interface Props {
   carPlatesNumber: ICar['number'] | undefined
@@ -11,7 +12,7 @@ export default function CarPlatesNumber(props: Props) {
   return (
     <div className='car-plates'>
       <h5 className='car-plates__number'>
-        { props.carPlatesNumber }
+        { parseCarNumber(props.carPlatesNumber) }
       </h5>
     </div>
   )
