+++ src/routes/orderPage/index.tsx
@@ -1,14 +1,17 @@
 import React from 'react'
 
 import './styles.scss'
-import useOrderPageContainer from './useOrderPageContainer'
+import { useAppSelector } from '../../hooks/usePreTypedHook';
 import Aside from '../../components/aside'
 import Header from '../../components/header'
 import OrderConfirmed from '../../components/orderConfirmed';
 import OrderPageTabs from '../../components/orderPageTabs';
 
 export default function OrderPage() {
-  const orderPageContainer = useOrderPageContainer()
+  const state = useAppSelector((state) => state)
+
+  const orderNumber = state.order.orderNumber
+  const isOrderConfirmed = state.order.isOrderConfirmed
 
   return (
     <main className='order-page'>
@@ -21,13 +24,13 @@ export default function OrderPage() {
           <Header />
         </div>
 
-        {!orderPageContainer.isOrderConfirmed
+        {!isOrderConfirmed
           ? <OrderPageTabs />
           : (
             <div className='order-page__header-row'>
               <div className='order-page__header-row--wrapper'>
                 <h6 className='order-page__header-row--title container'>
-                  Заказ номер { orderPageContainer.orderNumber }
+                  Заказ номер { orderNumber }
                 </h6>
               </div>
