--- src/routes/orderPage/useOrderPageContainer.tsx
@@ -1,13 +1,0 @@
-import { useAppSelector } from '../../hooks/usePreTypedHook';
-
-export default function useOrderPageContainer() {
-  const state = useAppSelector((state) => state)
-
-  const orderNumber = state.order.orderNumber
-  const isOrderConfirmed = state.order.isOrderConfirmed
-
-  return {
-    orderNumber,
-    isOrderConfirmed
-  }
-}
